Refactor all networking categories to be on CKClient
This was a bit more complicated than it sounds. It involved
getting rid of weird shared/current instances, changing the test
client a bit, etc., but ended up cleaning up everything. It also
meant getting rid of the CKILocalUser singleton instance and instead
making a readonly currentUser property on the CKIClient itself. Since
that was pretty much the only thin gin the CKILocalUser model besides
checking to see if we were logged in, I deleted the CKILocalUser and
got moved that method to the client as well.