Rename a bunch of OAuth 2 related stuff.
Why? To be a bit more consistent with the names of properties the
Canvas API gives us. For example, what we were calling oauthToken
is returned to us as accessToken. There was no real reason to call
this oauthToken, and it is also unclear whether this should be
oauthToken or oAuthToken, etc. Kills two birds with one stone.