better keychain key names.
The constants were some ending with name, others not. This was
inconsitent. Also confusing because `CKIKeychainCurrentUserName`
made the key look like it was only storing a user name, when in
actuality 'name' referred to the fact that the constant was the
name/key value in the keychain where the current user was stored.